Moved Setup Domain Flows into a single file
